server/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
cssselect2==0.7.0
decorator==5.1.1
distlib==0.3.8
Django==4.2.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.53.1
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-vision==3.7.3
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.22.0
identify==2.6.1
idna==3.7
import-ipynb==0.1.4
ipython==8.26.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nbformat==5.10.4
nodeenv==1.9.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.11.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyphen==0.16.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
rsa==4.9
setuptools==75.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
tinycss2==1.3.0
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.5
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3
